Refactor code to keep render.js modular and clean
Moved toggling project open/close state into index.js
Render.js will only RENDER those changes, not update the project
properties as well.

Also removed double parameters of functions that used to require
both the node and the object the node is rendering.

Did this by attaching the object to the node as a property.
And a lot of refactoring of functions that used 2 parameters.