Open and close projects without rerendering tab
Previously, the whole project+child todos node was rerendered to open and close.

Now there is a separate function to only remove or render the child todos.

This means less unnecessary rerendering.

Also this is good because the function is used both when creating project
and child todo for the first time and in this open/close function.
Less repeated code!